chore: move SvgIcon/Icons -> Icons/
