+++ packages/components/src/Banner/Banner.tsx
@@ -1,13 +1,13 @@
 import Heading, { HeadingElement } from "../Heading";
 import Button from "../Button";
-import CheckCircleIcon from "../SvgIcon/Icons/CheckCircle";
-import CloseIcon from "../SvgIcon/Icons/Close";
-import ForumIcon from "../SvgIcon/Icons/Forum";
-import NotificationsIcon from "../SvgIcon/Icons/Notifications";
+import CheckCircleIcon from "../Icons/CheckCircle";
+import CloseIcon from "../Icons/Close";
+import ForumIcon from "../Icons/Forum";
+import NotificationsIcon from "../Icons/Notifications";
 import React from "react";
-import SentimentVeryDissatisfiedIcon from "../SvgIcon/Icons/SentimentVeryDissatisfied";
+import SentimentVeryDissatisfiedIcon from "../Icons/SentimentVeryDissatisfied";
 import Text from "../Text";
-import WarningIcon from "../SvgIcon/Icons/Warning";
+import WarningIcon from "../Icons/Warning";
 import clsx from "clsx";
 import styles from "./Banner.module.css";
