+++ packages/components/src/SvgIcon/SvgIcon.tsx
@@ -57,7 +57,7 @@ interface SvgStyle extends React.CSSProperties {
 /**
  * Render arbitrary SVG path data while enforcing good accessibility practices.
  *
- * If you're looking for specific icon components, look in the `./Icons` directory.
+ * If you're looking for specific icon components, look in the `../Icons` directory.
  */
 function SvgIcon(props: SvgIconProps) {
   const {
